+++ packages/wayfinder-extension/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ar.io/wayfinder-extension
 
+## 1.0.2-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [86bdc2f]
+  - @ar.io/wayfinder-core@1.0.3-alpha.0
+
 ## 1.0.1
 
 ### Patch Changes